+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 ### Fixes
 - Shell escape pane titles to fix multi word and special character titles
+- Replace exists? with exist? for Ruby >= 3.2
 
 ## 3.1.2
 ### Fixes